backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.8
paddleocr==2.8.1
opencv-python==4.8.1.78
Pillow==11.3.0
numpy<2.0.0